// src/lib/generated/prisma/models/GameList.ts
export type GameListCountOutputTypeDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	/**
	 * Select specific fields to fetch from the GameListCountOutputType
	 */
	select?: Prisma.GameListCountOutputTypeSelect<ExtArgs> | null;
};
// ...
export type GameListSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<
	{
		id?: boolean;
		userId?: boolean;
		type?: boolean;
		displayMode?: boolean;
		tierLabels?: boolean;
		tierColors?: boolean;
		name?: boolean;
		slug?: boolean;
		description?: boolean;
		image?: boolean;
		background?: boolean;
		color?: boolean;
		accentColor?: boolean;
		privacy?: boolean;
		commentsHidden?: boolean;
		createdAt?: boolean;
		updatedAt?: boolean;
		entries?: boolean | Prisma.GameList$entriesArgs<ExtArgs>;
		user?: boolean | Prisma.UserDefaultArgs<ExtArgs>;
		likes?: boolean | Prisma.GameList$likesArgs<ExtArgs>;
		_count?: boolean | Prisma.GameListCountOutputTypeDefaultArgs<ExtArgs>;
	},
	ExtArgs["result"]["gameList"]
>;
// ...
export type GameListOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<
	| "id"
	| "userId"
	| "type"
	| "displayMode"
	| "tierLabels"
	| "tierColors"
	| "name"
	| "slug"
	| "description"
	| "image"
	| "background"
	| "color"
	| "accentColor"
	| "privacy"
	| "commentsHidden"
	| "createdAt"
	| "updatedAt",
	ExtArgs["result"]["gameList"]
>;
export type GameListInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	entries?: boolean | Prisma.GameList$entriesArgs<ExtArgs>;
	user?: boolean | Prisma.UserDefaultArgs<ExtArgs>;
	likes?: boolean | Prisma.GameList$likesArgs<ExtArgs>;
	_count?: boolean | Prisma.GameListCountOutputTypeDefaultArgs<ExtArgs>;
};
export type GameListIncludeCreateManyAndReturn<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	user?: boolean | Prisma.UserDefaultArgs<ExtArgs>;
};
export type GameListIncludeUpdateManyAndReturn<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	user?: boolean | Prisma.UserDefaultArgs<ExtArgs>;
};

export type $GameListPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	name: "GameList";
	objects: {
		entries: Prisma.$GameListEntryPayload<ExtArgs>[];
		user: Prisma.$UserPayload<ExtArgs>;
		likes: Prisma.$LikePayload<ExtArgs>[];
	};
	scalars: runtime.Types.Extensions.GetPayloadResult<
		{
			id: string;
			userId: string;
			type: $Enums.GameListType;
			displayMode: string;
			tierLabels: string[];
			tierColors: string[];
			name: string;
			slug: string | null;
			description: string | null;
			image: string | null;
			background: string | null;
			color: string | null;
			accentColor: string | null;
			privacy: string;
			commentsHidden: boolean;
			createdAt: Date;
			updatedAt: Date;
		},
		ExtArgs["result"]["gameList"]
	>;
	composites: {};
};

export type GameListGetPayload<S extends boolean | null | undefined | GameListDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$GameListPayload, S>;
// ...
export type GameList$entriesArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	/**
	 * Select specific fields to fetch from the GameListEntry
	 */
	select?: Prisma.GameListEntrySelect<ExtArgs> | null;
	/**
	 * Omit specific fields from the GameListEntry
	 */
	omit?: Prisma.GameListEntryOmit<ExtArgs> | null;
	/**
	 * Choose, which related nodes to fetch as well
	 */
	include?: Prisma.GameListEntryInclude<ExtArgs> | null;
	where?: Prisma.GameListEntryWhereInput;
	orderBy?: Prisma.GameListEntryOrderByWithRelationInput | Prisma.GameListEntryOrderByWithRelationInput[];
	cursor?: Prisma.GameListEntryWhereUniqueInput;
	take?: number;
	skip?: number;
	distinct?: Prisma.GameListEntryScalarFieldEnum | Prisma.GameListEntryScalarFieldEnum[];
};

/**
 * GameList.likes
 */
export type GameList$likesArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	/**
	 * Select specific fields to fetch from the Like
	 */
	select?: Prisma.LikeSelect<ExtArgs> | null;
	/**
	 * Omit specific fields from the Like
	 */
	omit?: Prisma.LikeOmit<ExtArgs> | null;
	/**
	 * Choose, which related nodes to fetch as well
	 */
	include?: Prisma.LikeInclude<ExtArgs> | null;
	where?: Prisma.LikeWhereInput;
	orderBy?: Prisma.LikeOrderByWithRelationInput | Prisma.LikeOrderByWithRelationInput[];
	cursor?: Prisma.LikeWhereUniqueInput;
	take?: number;
	skip?: number;
	distinct?: Prisma.LikeScalarFieldEnum | Prisma.LikeScalarFieldEnum[];
};

/**
 * GameList without action
 */
export type GameListDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
	/**
	 * Select specific fields to fetch from the GameList
	 */
	select?: Prisma.GameListSelect<ExtArgs> | null;
	/**
	 * Omit specific fields from the GameList
	 */
	omit?: Prisma.GameListOmit<ExtArgs> | null;
	/**
	 * Choose, which related nodes to fetch as well
	 */
	include?: Prisma.GameListInclude<ExtArgs> | null;
};