// src/lib/generated/prisma/models/UserGameEntry.ts
export type UserGameEntrySelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  userId?: boolean
  gameId?: boolean
  status?: boolean
  rating?: boolean
  timePlayed?: boolean
  timeMode?: boolean
  timeFinished?: boolean
  timeMastered?: boolean
  notes?: boolean
  favorite?: boolean
  addedAt?: boolean
  startedAt?: boolean
  finishedAt?: boolean
  masteredAt?: boolean
  createdAt?: boolean
  updatedAt?: boolean
  logs?: boolean | Prisma.UserGameEntry$logsArgs<ExtArgs>
  tags?: boolean | Prisma.UserGameEntry$tagsArgs<ExtArgs>
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
  game?: boolean | Prisma.GameDefaultArgs<ExtArgs>
  _count?: boolean | Prisma.UserGameEntryCountOutputTypeDefaultArgs<ExtArgs>
}, ExtArgs["result"]["userGameEntry"]>
// ...
export type UserGameEntryOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "userId" | "gameId" | "status" | "rating" | "timePlayed" | "timeMode" | "timeFinished" | "timeMastered" | "notes" | "favorite" | "addedAt" | "startedAt" | "finishedAt" | "masteredAt" | "createdAt" | "updatedAt", ExtArgs["result"]["userGameEntry"]>
export type UserGameEntryInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  logs?: boolean | Prisma.UserGameEntry$logsArgs<ExtArgs>
  tags?: boolean | Prisma.UserGameEntry$tagsArgs<ExtArgs>
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
  game?: boolean | Prisma.GameDefaultArgs<ExtArgs>
  _count?: boolean | Prisma.UserGameEntryCountOutputTypeDefaultArgs<ExtArgs>
}
// ...
export type $UserGameEntryPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "UserGameEntry"
  objects: {
    logs: Prisma.$UserGamePlayLogPayload<ExtArgs>[]
    tags: Prisma.$UserGameEntryTagPayload<ExtArgs>[]
    user: Prisma.$UserPayload<ExtArgs>
    game: Prisma.$GamePayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    gameId: number
    status: $Enums.GameStatus
    rating: number | null
    timePlayed: number | null
    timeMode: string
    timeFinished: number | null
    timeMastered: number | null
    notes: string | null
    favorite: boolean
    addedAt: Date
    startedAt: Date | null
    finishedAt: Date | null
    masteredAt: Date | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["userGameEntry"]>
  composites: {}
}

export type UserGameEntryGetPayload<S extends boolean | null | undefined | UserGameEntryDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$UserGameEntryPayload, S>
// ...
export type UserGameEntryDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the UserGameEntry
   */
  select?: Prisma.UserGameEntrySelect<ExtArgs> | null
  /**
   * Omit specific fields from the UserGameEntry
   */
  omit?: Prisma.UserGameEntryOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.UserGameEntryInclude<ExtArgs> | null
}